very minimal structure to decide if follower / leader or error on both
