+++ libs/lb/lb-rs/src/proxy_lb.rs
@@ -1,3 +1,5 @@
+use std::net::SocketAddrV4;
+
 pub struct ProxyLb {
-    
+    pub addr: SocketAddrV4
 }
